relocate modules
