+++ src/app/modules/directions/directions-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { DirectionsComponent } from './directions.component';
+import { DirectionsComponent } from './components/directions/directions.component';
 
 const routes: Routes = [{ path: '', component: DirectionsComponent }];
