+++ src/app/modules/directions/directions.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { DirectionsRoutingModule } from './directions-routing.module';
-import { DirectionsComponent } from './directions.component';
+import { DirectionsComponent } from './components/directions/directions.component';
 
 
 @NgModule({
