+++ src/app/modules/profile/profile-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { ProfileComponent } from './profile.component';
+import { ProfileComponent } from './components/profile/profile.component';
 
 const routes: Routes = [{ path: '', component: ProfileComponent }];
