+++ src/app/modules/profile/profile.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { ProfileRoutingModule } from './profile-routing.module';
-import { ProfileComponent } from './profile.component';
-import { TaskStatusPipe } from '../../pipes/status.pipe';
-import { RoundPipe } from '../../pipes/round.pipe';
+import { ProfileComponent } from './components/profile/profile.component';
+import { TaskStatusPipe } from './pipes/status.pipe';
+import { RoundPipe } from './pipes/round.pipe';
 
 @NgModule({
   declarations: [
